+++ packages/editor-preview/src/browser/editor-preview-contribution.ts
@@ -14,11 +14,12 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-import { ApplicationShell, CommonCommands, KeybindingContribution, KeybindingRegistry, SHELL_TABBAR_CONTEXT_MENU, Widget } from '@theia/core/lib/browser';
+import { ApplicationShell, CommonCommands, KeybindingContribution, KeybindingRegistry, SHELL_TABBAR_CONTEXT_PIN, Widget } from '@theia/core/lib/browser';
 import { nls } from '@theia/core/lib/common/nls';
 import { Command, CommandContribution, CommandRegistry, MenuContribution, MenuModelRegistry } from '@theia/core/lib/common';
 import { inject, injectable } from '@theia/core/shared/inversify';
 import { EditorPreviewWidget } from './editor-preview-widget';
+import { CurrentWidgetCommandAdapter } from '@theia/core/lib/browser/shell/current-widget-command-adapter';
 
 export namespace EditorPreviewCommands {
     export const PIN_PREVIEW_COMMAND = Command.toDefaultLocalizedCommand({
@@ -33,23 +34,16 @@ export class EditorPreviewContribution implements CommandContribution, MenuContr
     @inject(ApplicationShell) protected readonly shell: ApplicationShell;
 
     registerCommands(registry: CommandRegistry): void {
-        registry.registerCommand(EditorPreviewCommands.PIN_PREVIEW_COMMAND, {
-            execute: async (event?: Event) => {
-                const widget = this.getTargetWidget(event);
-                if (widget instanceof EditorPreviewWidget) {
-                    widget.convertToNonPreview();
-                    await this.shell.activateWidget(widget.id);
+        registry.registerCommand(EditorPreviewCommands.PIN_PREVIEW_COMMAND, new CurrentWidgetCommandAdapter(this.shell, {
+            execute: async title => {
+                if (title?.owner instanceof EditorPreviewWidget) {
+                    title.owner.convertToNonPreview();
+                    await this.shell.activateWidget(title.owner.id);
                 }
             },
-            isEnabled: (event?: Event) => {
-                const widget = this.getTargetWidget(event);
-                return widget instanceof EditorPreviewWidget && widget.isPreview;
-            },
-            isVisible: (event?: Event) => {
-                const widget = this.getTargetWidget(event);
-                return widget instanceof EditorPreviewWidget;
-            }
-        });
+            isEnabled: title => title?.owner instanceof EditorPreviewWidget && title.owner.isPreview,
+            isVisible: title => title?.owner instanceof EditorPreviewWidget,
+        }));
     }
 
     registerKeybindings(registry: KeybindingRegistry): void {
@@ -60,7 +54,7 @@ export class EditorPreviewContribution implements CommandContribution, MenuContr
     }
 
     registerMenus(registry: MenuModelRegistry): void {
-        registry.registerMenuAction(SHELL_TABBAR_CONTEXT_MENU, {
+        registry.registerMenuAction(SHELL_TABBAR_CONTEXT_PIN, {
             commandId: EditorPreviewCommands.PIN_PREVIEW_COMMAND.id,
             label: nls.localizeByDefault('Keep Open'),
             order: '6',